requirements.txt
chainlit>=1.0.0
litellm>=1.28.0
python-dotenv>=1.0.1